Add ability to use Docker for Node.js and Angular
Problem: This course is essentially a monorepo, so I needed to allow
Docker support for multiple processes; the Node.js server and the
Angular client (with simple static server).

Solution: Created two Dockerfiles, one for the client, one for the
server.  Documented in the Wiki regarding how to use it (Udemy: Angular
Testing).